Restructure repository to prepare for dotnet samples
