+++ samples/python/extraction/README.md
@@ -1,0 +1,13 @@
+# Extraction Samples
+
+This folder contains samples that demonstrate how to extract structured data from documents using Azure AI services.
+
+> [!NOTE]
+> All data extraction samples provide both an accuracy and confidence score for the extracted data. The accuracy score is calculated based on the similarity between the extracted data and the ground truth data. The confidence score is calculated based on OCR analysis confidence and `logprobs` in Azure OpenAI requests.
+
+| Sample                                                                                                                                                         | Description                                                                                                                                                                                               | Example Use Cases                                                                                                                     |
+| -------------------------------------------------------------------------------------------------------------------------------------------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------- |
+| [Data Extraction - Azure AI Document Intelligence + Azure OpenAI GPT-4o](./text/document-extraction-gpt.ipynb)                                                 | Demonstrates how to use Azure AI Document Intelligence pre-built layout and Azure OpenAI GPT models to extract structured data from documents.                                                            | Predominantly text-based documents such as invoices, receipts, and forms.                                                             |
+| [Data Extraction - Azure AI Document Intelligence + Phi](./text/document-extraction-phi.ipynb)                                                                 | Demonstrates how to use Azure AI Document Intelligence pre-built layout and Microsoft's Phi small language models to extract structured data from documents.                                              | Predominantly text-based documents such as invoices, receipts, and forms.                                                             |
+| [Data Extraction - Azure OpenAI GPT-4o with Vision](./vision/document-extraction-gpt-vision.ipynb)                                                             | Demonstrates how to use Azure OpenAI GPT-4o and GPT-4o-mini models to extract structured data from documents using their built-in vision capabilities.                                                    | Complex documents with a mix of text and images, including diagrams, signatures, selection marks, etc. such as reports and contracts. |
+| [Data Extraction - Comprehensive Azure AI Document Intelligence + Azure OpenAI GPT-4o with Vision](./multimodal/document-extraction-gpt-text-and-vision.ipynb) | Demonstrates how to improve the accuracy and confidence in extracting structured data from documents by combining Azure AI Document Intelligence and Azure OpenAI GPT-4o models with vision capabilities. | Any structured or unstructured document type.                                                                                         |
