+++ samples/python/modules/samples/models/redaction.py
