+++ samples/python/redaction/README.md
@@ -1,0 +1,8 @@
+# Redaction Samples
+
+This folder contains samples that demonstrate how to redact sensitive information from documents using Azure AI services.
+
+| Sample                                                                                                            | Description                                                                                                                                                                                                         | Example Use Cases                                                                                                                                                             |
+| ----------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| [Redaction - Azure AI Language PII Native Document Analysis](./document-redaction-language-native-document.ipynb) | Demonstrates how to use Azure AI Language Native Document Analysis features to redact personally identifiable information (PII) from documents.                                                                     | Redacting sensitive information from documents, such as names, addresses, and phone numbers.                                                                                  |
+| [Redaction - Azure AI Document Intelligence + Azure OpenAI GPT-4o](./document-redaction-gpt.ipynb)                | Demonstrates how to use Azure AI Document Intelligence pre-built layout and Azure OpenAI GPT models to redact sensitive information from documents using natural language instruction to determine redaction areas. | Need for more complex redaction rules, such as redacting based on context or relationships. Also works well for redacting PII, including names, addresses, and phone numbers. |
